Revert "chore: code to write common models to postgres (#871)"
This reverts commit 208f62953b78454f8f052e319e4ccb94f0f4d314.
