+++ openapi/crm/openapi.bundle.json
@@ -3161,12 +3161,12 @@
               "nullable": true,
               "example": "CA"
             },
-            "street_1": {
+            "street1": {
               "type": "string",
               "nullable": true,
               "example": "525 Brannan"
             },
-            "street_2": {
+            "street2": {
               "type": "string",
               "nullable": true,
               "example": null
@@ -3178,8 +3178,7 @@
             "country",
             "postal_code",
             "state",
-            "street_1",
-            "street_2"
+            "street1"
           ]
         },
         "example": [
@@ -3189,8 +3188,8 @@
             "country": "US",
             "postal_code": "94107",
             "state": "CA",
-            "street_1": "525 Brannan",
-            "street_2": null
+            "street1": "525 Brannan",
+            "street2": null
           }
         ]
       },
