--- packages/core/keys/account.ts
@@ -1,25 +1,0 @@
-import { SnakecasedKeysAccountWithTenant } from '@supaglue/types/crm';
-import { arrayOfAllKeys } from './util';
-
-export const keysOfSnakecasedAccountWithTenant = arrayOfAllKeys<SnakecasedKeysAccountWithTenant>()([
-  'provider_name',
-  'customer_id',
-  'id',
-  'remote_id',
-  'remote_created_at',
-  'remote_updated_at',
-  'remote_was_deleted',
-  'last_modified_at',
-  'name',
-  'description',
-  'industry',
-  'website',
-  'number_of_employees',
-  'addresses',
-  'phone_numbers',
-  'last_activity_at',
-  'lifecycle_stage',
-  'owner_id',
-  'owner',
-  'raw_data',
-]);
