+++ packages/core/mappers/crm/phone_number.ts
@@ -1,7 +1,6 @@
-import { PhoneNumber } from '@supaglue/types/base';
-import { SnakecasedKeysPhoneNumber } from '@supaglue/types/crm/common/phone_number';
+import { PhoneNumber } from '@supaglue/types/crm/common';
 
-export const toSnakecasedKeysPhoneNumber = (phoneNumber: PhoneNumber): SnakecasedKeysPhoneNumber => {
+export const toSnakecasedKeysPhoneNumber = (phoneNumber: PhoneNumber) => {
   return {
     phone_number: phoneNumber.phoneNumber,
     phone_number_type: phoneNumber.phoneNumberType,
