+++ packages/core/remotes/crm/salesforce/mappers.ts
@@ -1,5 +1,4 @@
-import type { Address, EmailAddress, PhoneNumber } from '@supaglue/types/base';
-import type {
+import {
   OpportunityStatus,
   RemoteAccount,
   RemoteAccountCreateParams,
@@ -15,6 +14,7 @@ import type {
   RemoteOpportunityUpdateParams,
   RemoteUser,
 } from '@supaglue/types/crm';
+import { Address, EmailAddress, PhoneNumber } from '@supaglue/types/crm/common';
 
 const CONTACT_ID_PREFIX = '003';
 const LEAD_ID_PREFIX = '00Q';
