+++ packages/core/services/destination_service.ts
@@ -1,7 +1,5 @@
 import type { PrismaClient } from '@supaglue/db';
 import type { Destination, DestinationCreateParams, DestinationUpdateParams } from '@supaglue/types';
-import { DestinationWriter } from '../destination_writers/base';
-import { PostgresDestinationWriter } from '../destination_writers/postgres';
 import { fromDestinationModel } from '../mappers/destination';
 
 export class DestinationService {
@@ -46,22 +44,5 @@ export class DestinationService {
       },
     });
     return fromDestinationModel(model);
-  }
-
-  public async getWriterByApplicationId(applicationId: string): Promise<DestinationWriter> {
-    const destinations = await this.getDestinationsByApplicationId(applicationId);
-    if (destinations.length === 0) {
-      throw new Error('No destinations found');
-    }
-    if (destinations.length > 1) {
-      throw new Error('Multiple destinations found');
-    }
-    const destination = destinations[0];
-    switch (destination.type) {
-      case 's3':
-        throw new Error('S3 destination not implemented');
-      case 'postgres':
-        return new PostgresDestinationWriter(destination);
-    }
   }
 }
