--- packages/types/base.ts
@@ -1,29 +1,0 @@
-export type Address = {
-  street1: string | null;
-  street2: string | null;
-  city: string | null;
-  state: string | null;
-  postalCode: string | null;
-  country: string | null;
-  addressType: 'primary' | 'billing' | 'mailing' | 'other' | 'shipping';
-};
-
-export type PhoneNumber = {
-  phoneNumber: string | null;
-  phoneNumberType: 'primary' | 'mobile' | 'fax';
-};
-
-export type EmailAddress = {
-  emailAddress: string;
-  emailAddressType: 'primary' | 'work';
-};
-
-export type LifecycleStage =
-  | 'subscriber'
-  | 'lead'
-  | 'marketingqualifiedlead'
-  | 'salesqualifiedlead'
-  | 'opportunity'
-  | 'customer'
-  | 'evangelist'
-  | 'other';
