+++ packages/types/crm/account.ts
@@ -1,14 +1,6 @@
 import type { BaseCrmModel, BaseCrmModelNonRemoteParams, BaseCrmModelRemoteOnlyParams, CustomFields, User } from '.';
-import { Address, LifecycleStage, PhoneNumber } from '../base';
 import { EqualsFilter, Filter } from '../filter';
-import { SnakecasedKeys } from '../snakecased_keys';
-
-export type SnakecasedKeysAccount = SnakecasedKeys<Account>;
-
-export type SnakecasedKeysAccountWithTenant = SnakecasedKeysAccount & {
-  provider_name: string;
-  customer_id: string;
-};
+import { Address, LifecycleStage, PhoneNumber } from './common';
 
 type BaseAccount = BaseCrmModel & {
   name: string | null;
