+++ packages/types/crm/common/address.ts
@@ -1,4 +1,9 @@
-import { Address } from '../../base';
-import { SnakecasedKeys } from '../../snakecased_keys';
-
-export type SnakecasedKeysAddress = SnakecasedKeys<Address>;
+export type Address = {
+  street1: string | null;
+  street2?: string | null;
+  city: string | null;
+  state: string | null;
+  postalCode: string | null;
+  country: string | null;
+  addressType: 'primary' | 'billing' | 'mailing' | 'other' | 'shipping';
+};
