+++ packages/types/crm/common/email_address.ts
@@ -1,4 +1,4 @@
-import { EmailAddress } from '../../base';
-import { SnakecasedKeys } from '../../snakecased_keys';
-
-export type SnakedcasedKeysEmailAddress = SnakecasedKeys<EmailAddress>;
+export type EmailAddress = {
+  emailAddress: string;
+  emailAddressType: 'primary' | 'work';
+};
