+++ packages/types/crm/common/phone_number.ts
@@ -1,4 +1,4 @@
-import { PhoneNumber } from '../../base';
-import { SnakecasedKeys } from '../../snakecased_keys';
-
-export type SnakecasedKeysPhoneNumber = SnakecasedKeys<PhoneNumber>;
+export type PhoneNumber = {
+  phoneNumber: string | null;
+  phoneNumberType: 'primary' | 'mobile' | 'fax';
+};
