+++ packages/types/crm/contact.ts
@@ -1,15 +1,7 @@
 import { BaseCrmModel, BaseCrmModelNonRemoteParams, BaseCrmModelRemoteOnlyParams, CustomFields, User } from '.';
-import { Address, EmailAddress, LifecycleStage, PhoneNumber } from '../base';
 import { EqualsFilter } from '../filter';
-import { SnakecasedKeys } from '../snakecased_keys';
 import type { Account } from './account';
-
-export type SnakecasedKeysContact = SnakecasedKeys<Contact>;
-
-export type SnakecasedKeysContactWithTenant = SnakecasedKeysContact & {
-  provider_name: string;
-  customer_id: string;
-};
+import { Address, EmailAddress, LifecycleStage, PhoneNumber } from './common';
 
 export type BaseContact = BaseCrmModel & {
   firstName: string | null;
