+++ packages/types/crm/index.ts
@@ -13,8 +13,8 @@ export const SUPPORTED_CRM_CONNECTIONS = [
 export type CRMProviderName = (typeof SUPPORTED_CRM_CONNECTIONS)[number];
 export type CRMProviderCategory = 'crm';
 
-export const CRM_COMMON_MODEL_TYPES = ['account', 'contact', 'lead', 'opportunity', 'user'] as const;
-export type CRMCommonModelType = (typeof CRM_COMMON_MODEL_TYPES)[number];
+export const CRM_COMMON_MODELS = ['account', 'contact', 'lead', 'opportunity', 'user'] as const;
+export type CRMCommonModelType = (typeof CRM_COMMON_MODELS)[number];
 
 export type CRMCommonModelTypeMap<T extends CRMCommonModelType> = {
   account: RemoteAccountTypes;
