+++ packages/types/crm/user.ts
@@ -1,12 +1,4 @@
 import { BaseCrmModel, BaseCrmModelNonRemoteParams, BaseCrmModelRemoteOnlyParams } from '.';
-import { SnakecasedKeys } from '../snakecased_keys';
-
-export type SnakecasedKeysUser = SnakecasedKeys<User>;
-
-export type SnakecasedKeysUserWithTenant = SnakecasedKeysUser & {
-  provider_name: string;
-  customer_id: string;
-};
 
 type BaseUser = BaseCrmModel & {
   name: string | null;
