+++ packages/types/engagement/index.ts
@@ -9,8 +9,8 @@ export const SUPPORTED_ENGAGEMENT_CONNECTIONS = ['outreach'] as const;
 export type EngagementProviderName = (typeof SUPPORTED_ENGAGEMENT_CONNECTIONS)[number];
 export type EngagementProviderCategory = 'engagement';
 
-export const ENGAGEMENT_COMMON_MODEL_TYPES = ['contact', 'user', 'sequence', 'mailbox', 'sequence_state'] as const;
-export type EngagementCommonModelType = (typeof ENGAGEMENT_COMMON_MODEL_TYPES)[number];
+export const ENGAGEMENT_COMMON_MODELS = ['contact', 'user', 'sequence', 'mailbox', 'sequence_state'] as const;
+export type EngagementCommonModelType = (typeof ENGAGEMENT_COMMON_MODELS)[number];
 
 export type EngagementCommonModelTypeMap<T extends EngagementCommonModelType> = {
   contact: RemoteContactTypes;
